Grid Reference changse
createMapGridPosition now accepts an easting (x) and northing (y) set of
coordinates and uses Marty's formatGrid to return a formatted gridref.
Fixed up the formatGrid to not redeclare the variable and changed the if
statements to be more efficient.
Sample now uses this system to report the grid ref, based on the receive
x and y positions.